move controller tests to folder
