+++ src/test/java/com/greengroupecommerce/controllers/HomeControllerTest.java
@@ -1,4 +1,4 @@
-package com.greengroupecommerce;
+package com.greengroupecommerce.controllers;
 
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
